feat: add news carousel and summary refactor
